Visualization extracted to different view; dedicated controller created to handle visualization-related issues.
