+++ src/main/resources/pl/edu/agh/iet/gg/meshgenerator/visualization/view/layout/MainWindowView.fxml
@@ -1,13 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
-<?import javafx.scene.Group?>
 <?import javafx.scene.layout.*?>
-<?import javafx.scene.SubScene?>
-<?import pl.edu.agh.iet.gg.meshgenerator.visualization.view.component.axis.XAxis?>
-<?import pl.edu.agh.iet.gg.meshgenerator.visualization.view.component.axis.YAxis?>
-<?import pl.edu.agh.iet.gg.meshgenerator.visualization.view.component.axis.ZAxis?>
-<?import pl.edu.agh.iet.gg.meshgenerator.visualization.view.component.RotatableGroup?>
 
 <GridPane xmlns="http://javafx.com/javafx/8.0.112-ea"
           xmlns:fx="http://javafx.com/fxml/1"
@@ -23,19 +17,7 @@
     </Pane>
 
     <Pane GridPane.columnIndex="1" GridPane.rowIndex="0">
-        <SubScene fx:id="graphScene" depthBuffer="true" height="600" width="640">
-            <root>
-                <RotatableGroup fx:id="environmentGroup">
-                    <Group>
-                        <XAxis/>
-                        <YAxis/>
-                        <ZAxis/>
-                    </Group>
-
-                    <Group fx:id="graphGroup"/>
-                </RotatableGroup>
-            </root>
-        </SubScene>
+        <fx:include fx:id="graphScene" source="GraphView.fxml"/>
     </Pane>
 
 </GridPane>
